TLDR-444 added word extraction from pdfminer; pdfminer refactoring
